requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
altgraph
annotated-types
anyio
asgiref
attrs
beautifulsoup4
blinker
certifi
cffi
charset-normalizer
click
contourpy
crispy-bootstrap5
cryptography
cycler
deprecation
distlib
dj-database-url
Django
django-bootstrap5
django-cors-headers
django-crispy-forms
django-email
django-environ
django-extensions
django-filter
django-humanize
django-widget-tweaks
djangorestframework
djangorestframework_simplejwt
drf-yasg
filelock
Flask
flatbuffers
fonttools
frozenlist
gotrue
gunicorn
h11
h2
hpack
httpcore
httpx
humanize
hyperframe
idna
inflection
itsdangerous
jax
jaxlib
Jinja2
joblib
kiwisolver
Markdown
MarkupSafe
ml_dtypes
multidict
mysqlclient
numpy
opencv-contrib-python
opencv-python
opt_einsum
packaging
pefile
pillow
pipenv
platformdirs
postgrest
propcache
protobuf
psycopg2
psycopg2-binary
PyAutoGUI
pycparser
pydantic
pydantic_core
pygame
PyGetWindow
pyinstaller
pyinstaller-hooks-contrib
PyJWT
Pympler
PyMsgBox
PyMySQL
pyparsing
PyPDF2
pyperclip
PyRect
PyScreeze
python-dateutil
pyttsx3
pytweening
pytz
pywhatkit
PyYAML
qrcode
realtime
requests
scikit-learn
scipy
sentencepiece
setuptools
six
sniffio
sounddevice
soupsieve
sqlparse
sslcommerz-lib
storage3
StrEnum
supabase
supafunc
threadpoolctl
typing_extensions
tzdata
uritemplate
urllib3
virtualenv
websockets
Werkzeug
wheel
whitenoise
wikipedia
yarl